Move noship dir to image instead of www
This is an artifict needed by the swupd-server when creating update content
in the webdir, it should not exist in the webdir since it is being created
as part of the chroot building process. All content produced by the chroot
builder should live in the STATE_DIR/image/VER directory.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>