Don't use os-core chroot as base for other bundles
Currently there is an issue with running yum in docker where the yum db
would become corrupt when attempting to reuse an existing db file under
certain conditons that are not well understood.

In order to avoid this issue, instead of using a copy of the os-core
chroot when building the other chroots, create all chroots from fresh
instead.

This change does have a negative performance impact since yum will be
duplicating installation steps but this change was the most reliable way
to create chroots under docker.

As an aside, skipping the yum install of filesystem as a pre-step was
also looked at but dismissed due to the possibility of permissions
getting set incorrectly when creating the chroots because of yum install
order.